include two binaries
