+++ README.md
@@ -3,6 +3,8 @@
 Utility to check Emeris data against actual blockchain nodes, in order to report
 inaccuracies.
 
+Includes two binaries `balcheck` (runs http server)  and `balcheck-cli` (contains client command)
+
 ## Checks
 
 Available checks are:
@@ -11,7 +13,7 @@ Available checks are:
 - staking balance
 - unstaking balances
 
-## Usage
+## `balcheck` usage
 
 ```sh
 make balcheck
@@ -29,4 +31,13 @@ Available endpoints are:
 - /check/:address
 
 Example: http://localhost:8081/check/cosmos1qymla9gh8z2cmrylt008hkre0gry6h92sxgazg
+
+## `balcheck-cli` usage
+
+```sh
+make balcheck-cli
+./build/balcheck-cli -addr cosmos1qymla9gh8z2cmrylt008hkre0gry6h92sxgazg
+```
+
+This command will launch the checks for all the chains available in Emeris.
