+++ api/account/api.go
@@ -7,9 +7,8 @@ import (
 
 	"github.com/damianopetrungaro/golog"
 	"github.com/emerishq/balcheck/pkg/bech32"
-	"github.com/emerishq/balcheck/pkg/checker"
 	"github.com/emerishq/balcheck/pkg/emeris"
-	"github.com/emerishq/balcheck/pkg/lcd"
+	"github.com/emerishq/balcheck/utils"
 	"github.com/gorilla/mux"
 )
 
@@ -34,59 +33,6 @@ func CheckAddress(emerisClient *emeris.Client, w *golog.BufWriter) func(http.Res
 
 		fmt.Fprint(response, "Started balance checking")
 
-		for _, chain := range chains {
-			go func(chain checker.Chain) {
-				defer w.Flush()
-				lcdClient := lcd.NewClient(chain)
-				log := golog.With(
-					golog.String("check", "balances"),
-					golog.String("chain", chain.Name),
-					golog.String("api_url", emerisClient.BalancesURL(addr)),
-					golog.String("lcd_url", lcdClient.BalancesURL(addr)),
-				)
-				log.Info(ctx, "started testing")
-
-				err := checker.BalanceCheck(ctx, addr, emerisClient.Balances, lcdClient.Balances)
-				if err != nil {
-					log.With(golog.Err(err)).Error(ctx, "balance mismatch")
-				}
-			}(chain)
-		}
-
-		for _, chain := range chains {
-			go func(chain checker.Chain) {
-				lcdClient := lcd.NewClient(chain)
-				log := golog.With(
-					golog.String("check", "staking balances"),
-					golog.String("chain", chain.Name),
-					golog.String("api_url", emerisClient.StakingBalancesURL(addr)),
-					golog.String("lcd_url", lcdClient.StakingBalancesURL(addr)),
-				)
-				log.Info(ctx, "started testing")
-
-				err := checker.BalanceCheck(ctx, addr, emerisClient.StakingBalances, lcdClient.StakingBalances)
-				if err != nil {
-					log.With(golog.Err(err)).Error(ctx, "staking balance mismatch")
-				}
-			}(chain)
-		}
-
-		for _, chain := range chains {
-			go func(chain checker.Chain) {
-				lcdClient := lcd.NewClient(chain)
-				log := golog.With(
-					golog.String("check", "unbonding balances"),
-					golog.String("chain", chain.Name),
-					golog.String("api_url", emerisClient.UnstakingBalancesURL(addr)),
-					golog.String("lcd_url", lcdClient.UnstakingBalancesURL(addr)),
-				)
-				log.Info(ctx, "started testing")
-
-				err := checker.BalanceCheck(ctx, addr, emerisClient.UnstakingBalances, lcdClient.UnstakingBalances)
-				if err != nil {
-					log.With(golog.Err(err)).Error(ctx, "unbonding balance mismatch")
-				}
-			}(chain)
-		}
+		utils.CheckBalances(emerisClient, chains, w, addr, false)
 	}
 }
